most server views

+++ FirmaKurierska.SharedKernel/Dto/OrderDto.cs
@@ -8,7 +8,7 @@ namespace FirmaKurierska.SharedKernel.Dto
         public int Id { get; set; }
         public OrderStatus Status { get; set; }
         public OrderSize Size { get; set; }
-        public decimal TotalCost { get; set; }
+        public double TotalCost { get; set; }
         public Address PickupLocation { get; set; }
         public Address Destination { get; set; }
         public DateTime ShippingDate { get; set; }
